Un-revert "Improve (and shorten) query polling implementation. (#4243)"
This reverts commit 9739ff6738925fc6851d0cdf81d8431abc8ba71e.

Now that we have a uniform interface for terminating ApolloClient
instances (#4336), there should be no need for any external code to access
the QueryScheduler abstraction, which this commit removes.

We should wait to merge and release this change until after
https://github.com/apollographql/react-apollo/pull/2741 has been merged
and released, so that we don't break older versions of MockedProvider.